LayoutTests/imported/w3c/web-platform-tests/editing/other/delete-without-unwrapping-first-line-of-child-block_method=BackspaceKey&lineBreak=br-expected.txt


ef
ghi

FAIL Backspace at abc<br>{<br><div id="child">]def<br>ghi</div> assert_in_array: Backspace at abc<br>{<br><div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> value "abc<br>def<br><div id=\"child\">ghi</div>" not in array ["abc<br><div id=\"child\">def<br>ghi</div>", "abc<div id=\"child\">def<br>ghi</div>"]
FAIL Typing "X" at abc<br>{<br><div id="child">]def<br>ghi</div> assert_in_array: Typing "X" at abc<br>{<br><div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> value "abc<br>Xdef<br><div id=\"child\">ghi</div>" not in array ["abc<br><div id=\"child\">Xdef<br>ghi</div>", "abc<div id=\"child\">Xdef<br>ghi</div>"]
PASS Backspace at <br>[abc<br><div id="child">]def<br>ghi</div>
PASS Typing "X" at <br>[abc<br><div id="child">]def<br>ghi</div>
FAIL Backspace at <br>{<br><div id="child">]def<br>ghi</div> assert_equals: Backspace at <br>{<br><div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<br><div id=\"child\">def<br>ghi</div>" but got "<br>def<br><div id=\"child\">ghi</div>"
FAIL Typing "X" at <br>{<br><div id="child">]def<br>ghi</div> assert_equals: Typing "X" at <br>{<br><div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> expected "<br><div id=\"child\">Xdef<br>ghi</div>" but got "<br>Xdef<br><div id=\"child\">ghi</div>"
FAIL Backspace at {<br><br><div id="child">]def<br>ghi</div> assert_equals: Backspace at {<br><br><div id="child">]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "def<br><div id=\"child\">ghi</div>"
FAIL Typing "X" at {<br><br><div id="child">]def<br>ghi</div> assert_equals: Typing "X" at {<br><br><div id="child">]def<br>ghi</div> should not unwrap the first line of the child <div> expected "<div id=\"child\">Xdef<br>ghi</div>" but got "Xdef<br><div id=\"child\">ghi</div>"
PASS Backspace at [abc<br><br><div id="child">]def<br>ghi</div>
PASS Typing "X" at [abc<br><br><div id="child">]def<br>ghi</div>
FAIL Backspace at abc<br>{<br><div id="child">d]ef<br>ghi</div> assert_in_array: Backspace at abc<br>{<br><div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected text in the <div> value "abc<br>ef<br><div id=\"child\">ghi</div>" not in array ["abc<br><div id=\"child\">ef<br>ghi</div>", "abc<div id=\"child\">ef<br>ghi</div>"]
FAIL Typing "X" at abc<br>{<br><div id="child">d]ef<br>ghi</div> assert_in_array: Typing "X" at abc<br>{<br><div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> value "abc<br>Xef<br><div id=\"child\">ghi</div>" not in array ["abc<br><div id=\"child\">Xef<br>ghi</div>", "abc<div id=\"child\">Xef<br>ghi</div>"]
PASS Backspace at <br>[abc<br><div id="child">d]ef<br>ghi</div>
PASS Typing "X" at <br>[abc<br><div id="child">d]ef<br>ghi</div>
FAIL Backspace at <br>{<br><div id="child">d]ef<br>ghi</div> assert_equals: Backspace at <br>{<br><div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected content in the <div> expected "<br><div id=\"child\">ef<br>ghi</div>" but got "<br>ef<br><div id=\"child\">ghi</div>"
FAIL Typing "X" at <br>{<br><div id="child">d]ef<br>ghi</div> assert_equals: Typing "X" at <br>{<br><div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> expected "<br><div id=\"child\">Xef<br>ghi</div>" but got "<br>Xef<br><div id=\"child\">ghi</div>"
FAIL Backspace at {<br><br><div id="child">d]ef<br>ghi</div> assert_equals: Backspace at {<br><br><div id="child">d]ef<br>ghi</div> should delete only the preceding empty line of the child <div> and selected content in the <div> expected "<div id=\"child\">ef<br>ghi</div>" but got "ef<br><div id=\"child\">ghi</div>"
FAIL Typing "X" at {<br><br><div id="child">d]ef<br>ghi</div> assert_equals: Typing "X" at {<br><br><div id="child">d]ef<br>ghi</div> should not unwrap the first line of the child <div> expected "<div id=\"child\">Xef<br>ghi</div>" but got "Xef<br><div id=\"child\">ghi</div>"
PASS Backspace at [abc<br><br><div id="child">d]ef<br>ghi</div>
PASS Typing "X" at [abc<br><br><div id="child">d]ef<br>ghi</div>
FAIL Backspace at abc<br><br><div id="child">[]def<br>ghi</div> assert_in_array: Backspace at abc<br><br><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> value "abc<br><br><div id=\"child\">ef<br>ghi</div>" not in array ["abc<br><div id=\"child\">def<br>ghi</div>", "abc<div id=\"child\">def<br>ghi</div>"]
FAIL Backspace at <br><br><div id="child">[]def<br>ghi</div> assert_equals: Backspace at <br><br><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<br><div id=\"child\">def<br>ghi</div>" but got "<br><br><div id=\"child\">ef<br>ghi</div>"
FAIL Backspace at <br><div id="child">[]def<br>ghi</div> assert_equals: Backspace at <br><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "<br><div id=\"child\">ef<br>ghi</div>"
FAIL Backspace at <b>abc<br><br></b></b><div id="child">[]def<br>ghi</div> assert_in_array: Backspace at <b>abc<br><br></b></b><div id="child">[]def<br>ghi</div> should delete only the preceding empty line of the child <div> (<b> should stay) value "<b>abc<br><br></b><div id=\"child\">ef<br>ghi</div>" not in array ["<b>abc<br></b><div id=\"child\">def<br>ghi</div>", "<b>abc</b><div id=\"child\">def<br>ghi</div>"]
FAIL Backspace at <b><br></b><div id="child">[]def<br>ghi</div> assert_equals: Backspace at <b><br></b><div id="child">[]def<br>ghi</div> should delete only the preceding empty line (including the <b>) of the child <div> expected "<div id=\"child\">def<br>ghi</div>" but got "<b><br></b><div id=\"child\">ef<br>ghi</div>"
FAIL getTargetRanges() for Backspace at <br>[abc<br><div id="child">]def<br>ghi</div> should return a range before the child <div> EditorTestUtils.getRangeArrayDescription is not a function. (In 'EditorTestUtils.getRangeArrayDescription(expectedTargetRanges)', 'EditorTestUtils.getRangeArrayDescription' is undefined)
FAIL getTargetRanges() for Typing "X" at <br>[abc<br><div id="child">]def<br>ghi</div> should return a range ending in the child <div> EditorTestUtils.getRangeArrayDescription is not a function. (In 'EditorTestUtils.getRangeArrayDescription(expectedTargetRanges)', 'EditorTestUtils.getRangeArrayDescription' is undefined)
FAIL getTargetRanges() for Backspace at [abc<br><br><div id="child">]def<br>ghi</div> should return a range before the child <div> EditorTestUtils.getRangeArrayDescription is not a function. (In 'EditorTestUtils.getRangeArrayDescription(expectedTargetRanges)', 'EditorTestUtils.getRangeArrayDescription' is undefined)
FAIL getTargetRanges() for Typing "X" at [abc<br><br><div id="child">]def<br>ghi</div> should return a range ending in the child <div> EditorTestUtils.getRangeArrayDescription is not a function. (In 'EditorTestUtils.getRangeArrayDescription(expectedTargetRanges)', 'EditorTestUtils.getRangeArrayDescription' is undefined)
FAIL getTargetRanges() for Backspace at <br>[abc<br><div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div> EditorTestUtils.getRangeArrayDescription is not a function. (In 'EditorTestUtils.getRangeArrayDescription(expectedTargetRanges)', 'EditorTestUtils.getRangeArrayDescription' is undefined)
FAIL getTargetRanges() for Typing "X" at <br>[abc<br><div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div> EditorTestUtils.getRangeArrayDescription is not a function. (In 'EditorTestUtils.getRangeArrayDescription(expectedTargetRanges)', 'EditorTestUtils.getRangeArrayDescription' is undefined)
FAIL getTargetRanges() for Backspace at [abc<br><br><div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div> EditorTestUtils.getRangeArrayDescription is not a function. (In 'EditorTestUtils.getRangeArrayDescription(expectedTargetRanges)', 'EditorTestUtils.getRangeArrayDescription' is undefined)
FAIL getTargetRanges() for Typing "X" at [abc<br><br><div id="child">d]ef<br>ghi</div> should return a range ends at start of the child <div> EditorTestUtils.getRangeArrayDescription is not a function. (In 'EditorTestUtils.getRangeArrayDescription(expectedTargetRanges)', 'EditorTestUtils.getRangeArrayDescription' is undefined)
